// src/Modules/Dynamic/UserManagement/Models/IUserMaster.ts
export interface IUserMaster {
  userId: number;

  fullName: string;

  email: string;

  mobile: number;

  gender: string;

  date?: string;

  ssn: number;

  password?: string;

  accStatus: string;

  createdDate?: string;

  updatedDate?: string;

  createdBy?: string;

  updatedBy?: string;
}